Rename files standardviewmodels -> viewmodelfactory
