+++ examples/concurrentplot/concurrentplotcore/graphpropertywidget.cpp
@@ -10,7 +10,7 @@
 #include "graphpropertywidget.h"
 #include "graphmodel.h"
 #include <QBoxLayout>
-#include <mvvm/viewmodel/standardviewmodels.h>
+#include <mvvm/factories/viewmodelfactory.h>
 #include <mvvm/viewmodel/viewmodel.h>
 #include <mvvm/widgets/itemstreeview.h>
