+++ examples/plotgraphs/plotgraphscore/graphpropertywidget.cpp
@@ -11,7 +11,7 @@
 #include "graphmodel.h"
 #include <QBoxLayout>
 #include <QPushButton>
-#include <mvvm/viewmodel/standardviewmodels.h>
+#include <mvvm/factories/viewmodelfactory.h>
 #include <mvvm/viewmodel/viewmodel.h>
 #include <mvvm/widgets/itemstreeview.h>
