+++ source/libmvvm_viewmodel/mvvm/factories/CMakeLists.txt
@@ -2,4 +2,6 @@ target_sources(${library_name} PRIVATE
     viewmodelcontrollerbuilder.cpp
     viewmodelcontrollerbuilder.h
     viewmodelcontrollerfactory.h
+    viewmodelfactory.cpp
+    viewmodelfactory.h
 )
