+++ source/libmvvm_viewmodel/mvvm/factories/viewmodelfactory.cpp
@@ -11,7 +11,7 @@
 #include <mvvm/viewmodel/propertyflatviewmodel.h>
 #include <mvvm/viewmodel/propertytableviewmodel.h>
 #include <mvvm/viewmodel/propertyviewmodel.h>
-#include <mvvm/viewmodel/standardviewmodels.h>
+#include <mvvm/factories/viewmodelfactory.h>
 #include <mvvm/viewmodel/topitemsviewmodel.h>
 
 using namespace ModelView;
